refactor(debug): put tracer in separate file
